// src/logic/rest/api.ts
export interface Body1 {
    /**
     * First Name of the user
     * @type {string}
     * @memberof Body1
     */
    firstname: string;
    /**
     * Last Name of the user
     * @type {string}
     * @memberof Body1
     */
    lastname: string;
    /**
     * The username
     * @type {string}
     * @memberof Body1
     */
    username: string;
    /**
     * The e-mail of the user
     * @type {string}
     * @memberof Body1
     */
    email: string;
    /**
     * Password
     * @type {string}
     * @memberof Body1
     */
    password: string;
}
// ...
export interface User {
    /**
     * The id of the user
     * @type {number}
     * @memberof User
     */
    id: number;
    /**
     * The username
     * @type {string}
     * @memberof User
     */
    username: string;
    /**
     * First name of the user
     * @type {string}
     * @memberof User
     */
    firstname: string;
    /**
     * Last name of the user (only visible to admins)
     * @type {string}
     * @memberof User
     */
    lastname?: string;
    /**
     * E-mail of the user (only visible to admins)
     * @type {string}
     * @memberof User
     */
    email?: string;
    /**
     * `true` if the email is verified, else `false`
     * @type {boolean}
     * @memberof User
     */
    emailVerified: boolean;
}